requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pip==21.3.1
setuptools==60.2.0
Werkzeug==2.2.2
wheel==0.37.1
zipp==3.12.1

textblob~=0.17.1
nltk~=3.8.1
spacy~=3.5.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
